Fixed the bug in GameSystem so that the addEntity function is only called when the entity isnt already in the entities set.
Next step is to adjust physics collider for tank in y direction as it's a couple of pixels off, and to implement rotation and damping.

+++ app/src/main/java/com/example/helloworld/systems/physics/PhysicsBodyGenerator.java
@@ -8,7 +8,7 @@ import org.jbox2d.dynamics.BodyType;
 import org.jbox2d.dynamics.FixtureDef;
 
 public class PhysicsBodyGenerator {
-    public static PhysicsComponent createPhysicsBody(Vec2 position, boolean dynamic, int width, int height){
+    public static PhysicsComponent createPhysicsBody(Vec2 position, boolean dynamic, float width, float height){
         PhysicsComponent physicsComponent = new PhysicsComponent();
         //body definition
         BodyDef bd = new BodyDef();
